Display process name in error message on failure to create a tab
When a tab could not be created an error message included a tab name, but
not the process name, even though it did instruct to search for the
process in Task Manager.

To simplify the operation, the shell is now included in that error message.

This is an attempt to fix Improve Error Messaging - ID: 1799638