Remove accelerator from Command label on Hotkeys Settings page
Patch 2749871 "Bug Fix for 2250711" By Phillip Foose ( phillipfoose ):
This is a very simple bug fix, on the hot keys dialog, you had &C for
Command and &C for Clear. Sine the Command field is a read only text box,
I've removed the accelerator key, so that alt-c works as expected and
clears the key.