Use GetSettingsFileName to calculate absolute path of console.xml correctly
Patch 2749841 "Bug Fix for 1959580" by Phillip Foose ( phillipfoose ):
The full path to console.xml was updated in SettingsHandler only if a user
opened the Settings dialog, changed something and clicked OK. Otherwise,
the relative path - console.xml - was used.

When a user had "Save settings to user directory" checked,
%appdata%\console.xml was correctly loaded on startup, but the relative
path would be used, for example, for saving initial position on exit (see
Appearance -> Position group in the Settings). Hence, it appeared as
the Initial Position was never saved.

In addition to Phillip's change, another instances of possibly wrong path
to the settings file were fixed.